TSK-1576: Implemented tests to retrieve tasks for different user roles
