+++ lib/taskana-core/src/test/java/acceptance/task/TaskEngineAccTest.java
@@ -1,4 +1,4 @@
-package acceptance.security;
+package acceptance.task;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
